feat(rust): enhance orchestrator and execution system
Refactor unified workflow executor with resume capabilities.
Add error monitor and event system modules.
Enhance MCP integration and AI workflows.
Improve check executor analyzers and command builder.
Add step registry and execution core modules.
Update database schema and context handling.

+++ src-tauri/src/orchestrator/compression.rs
@@ -10,6 +10,8 @@
 //! - Summarization of old knowledge entries
 //! - Preservation of recent context
 
+#![allow(dead_code)]
+
 use serde::{Deserialize, Serialize};
 use std::sync::Arc;
 use tracing::{debug, info, warn};
@@ -413,7 +415,7 @@ impl CompressionService {
         category: &str,
         entries: &[StoredTaskKnowledge],
     ) -> Result<KnowledgeSummary, String> {
-        let covered_iterations: Vec<u32> = entries.iter().map(|e| e.iteration as u32).collect();
+        let covered_iterations: Vec<u32> = entries.iter().map(|e| e.iteration).collect();
 
         // Build a simple extractive summary
         // In a production system, this could use AI summarization
